face_attendance 2/requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dlib==20.0.0
dnspython==2.8.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
flask-cors==6.0.2
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
opencv-python==4.13.0.90
pillow==12.1.0
pycparser==2.23
PyJWT==2.10.1
pymongo==4.16.0
python-dotenv==1.2.1
Werkzeug==3.1.5
